# b/static/swagger.yaml
openapi: 3.0.0
info:
  title: Planner App API. 
  description: |
    API documnetation for the Flask Planner App backend.<br/>
    How to test the API:
      1) Use /sign-up to create an account
      2) Use /login to the login to account created
      3) You now have sufficient auth to test other endpoints (e.g., test /read-project, read-objetives and read-task to inspect the automatically created entities)
  version: 1.0.0
...
servers:
  - url: http://127.0.0.1:5000/api
    description: Testing server
  - url: https://plannerappbackendnc-dad9ghcgd6ame2cc.ukwest-01.azurewebsites.net/api
    description:  Main (production) server
paths:
...
  /create-task:
    post:
      tags: [Task]
      summary: Create a new task
      description: Create a new task within an objective
      security:
        - bespoke_session: []
        - session_AT: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                taskNumber:
                  type: integer
                  description: Optional task number
                status:
                  type: string
                  default: "To-Do"
                  enum: ["To-Do", "In Progress", "Paused", "Completed"]
                description:
                  type: string
                  description: Task description (character limit defined in env)
                durationEst:
                  type: integer
                  description: Estimated duration in minutes (required)
                duration:
                  type: integer
                  description: Actual duration in minutes
                priorityScore:
                  type: integer
                  default: 1
                  description: Priority score (higher = more important)
                scheduledStart:
                  type: string
                  format: date
                  description: Scheduled start date (YYYY-MM-DD)
                start:
                  type: string
                  format: date-time
                  description: Actual start time (YYYY-MM-DDTHH:MM)
                finish:
                  type: string
                  format: date-time
                  description: Actual finish time (YYYY-MM-DDTHH:MM)
                isRecurring:
                  type: boolean
                  default: false
                wasPaused:
                  type: boolean
                  default: false
                parentTaskId:
                  type: integer
                  description: ID of parent task if this is a subtask
                tag:
                  type: string
                objectiveId:
                  type: integer
                  description: ID of the objective this task belongs to (required)
              required:
                - description
                - durationEst
                - objectiveId
      responses:
        '201':
          description: |
            Success: Task created successfully. Possible messages:
            - "Success: Added Task to db!"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  task:
                    $ref: '#/components/schemas/Task'
        '400':
          description: |
            Bad request. Possible validation errors:
            - "Failure: Objective ID missing. The Task is not assigned any project."
            - "Failure: Task is missing a description."
            - "Failure: The description is over the {task_description_limit} char limit."
            - "Failure: Task is missing a duration_est (mins)."
            - "Failure: Task duration cannot be zero or negative."
            - "Failure: Task cannot have a finish date and be incomplete."
        '401':
          description: |
            Unauthorized. Possible authentication errors from decorators:
            - "Request is missing access token. Please login to refresh access token"
            - "Deserialisation or Signiture verificaiton of access token cookie has failed! Reason: {error details}"
            - "Invalid Access Token! Reason: {error details}"
            - "Failure: User is not logged in (no b_sc). Please login!"
            - "Failure: Could not decrypt the bespoke_session cookies. Reason: {error details}"
            - "Failure: Refresh token is invalid. Please login"
            - "Failure: Please login. Refresh token has expired."
        '403':
          description: |
            Forbidden. Possible messages:
            - "Failure: The specified objective does not belong to the user."
        '404':
          description: |
            Not found. Possible messages:
            - "Failure: The Objective ID provided does not exist."
            - "Failure: The Task could not be added to the db! Reason: {error details}"
            - "Failure: User is not logged in (no rt). Please login!"
  /read-tasks:
    get:
      tags: [Task]
      summary: Get all tasks
      description: Retrieve all tasks for the current user
      security:
        - bespoke_session: []
        - session_AT: []
      responses:
        '200':
          description: |
            Success: Tasks retrieved successfully. Possible messages:
            - "Success: user's tasks loaded"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  tasks:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
        '401':
          description: |
            Unauthorized. Possible authentication errors from decorators:
            - "Request is missing access token. Please login to refresh access token"
            - "Deserialisation or Signiture verificaiton of access token cookie has failed! Reason: {error details}"
            - "Invalid Access Token! Reason: {error details}"
            - "Failure: User is not logged in (no b_sc). Please login!"
            - "Failure: Could not decrypt the bespoke_session cookies. Reason: {error details}"
            - "Failure: Refresh token is invalid. Please login"
            - "Failure: Please login. Refresh token has expired."
        '404':
          description: |
            Not found. Possible messages:
            - "Failure: Could not read user task! Reason: {error details}"
            - "Failure: User is not logged in (no rt). Please login!"
  /query-tasks:
    get:
      tags: [Task]
      summary: Query tasks with filters
      description: Retrieve tasks with extensive filtering and pagination options
      security:
        - bespoke_session: []
        - session_AT: []
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
        - name: perPage
          in: query
          description: Items per page for pagination
          schema:
            type: integer
        - name: objectiveId
          in: query
          description: Filter by objective ID
          schema:
            type: integer
        - name: status
          in: query
          description: Filter by status (comma-separated for multiple values)
          schema:
            type: string
        - name: selectedDate
          in: query
          description: Filter by specific date (format YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: sitePage
          in: query
          description: Context for filtering (homepage, kanban, planner)
          schema:
            type: string
        - name: periodStart
          in: query
          description: Start date for period filter (format YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: periodEnd
          in: query
          description: End date for period filter (format YYYY-MM-DD)
          schema:
            type: string
            format: date
      responses:
        '200':
          description: |
            Success: Tasks retrieved successfully. Possible messages:
            - "Success: user's tasks loaded"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  tasks:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
                  taskObjectives:
                    type: array
                    items:
                      $ref: '#/components/schemas/Objective'
                  taskProjects:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
                  _pages:
                    type: integer
                  _currentPage:
                    type: integer
                  _perPage:
                    type: integer
                  _hasNext:
                    type: boolean
                  _hasPrev:
                    type: boolean
                  _itemCount:
                    type: integer
                  _objectiveId:
                    type: integer
                  _status:
                    type: string
                  _selectedDate:
                    type: string
                  periodStart:
                    type: string
                  periodEnd:
                    type: string
        '400':
          description: |
            Bad request. Possible messages:
            - "Failure: End date cannot be before start date. Please change this."
        '401':
          description: |
            Unauthorized. Possible authentication errors from decorators:
            - "Request is missing access token. Please login to refresh access token"
            - "Deserialisation or Signiture verificaiton of access token cookie has failed! Reason: {error details}"
            - "Invalid Access Token! Reason: {error details}"
            - "Failure: User is not logged in (no b_sc). Please login!"
            - "Failure: Could not decrypt the bespoke_session cookies. Reason: {error details}"
            - "Failure: Refresh token is invalid. Please login"
            - "Failure: Please login. Refresh token has expired."
        '404':
          description: |
            Not found. Possible messages:
            - "Failure: Could not read user task! Reason: {error details}"
            - "Failure: User is not logged in (no rt). Please login!"
  /get-tasks-objective-and-project/{task_id}:
    get:
      tags: [Task]
      summary: Get objective and project for a task
      description: Retrieve the objective and project associated with a specific task
      security:
        - bespoke_session: []
        - session_AT: []
      parameters:
        - name: task_id
          in: path
          required: true
          description: ID of the task
          schema:
            type: integer
      responses:
        '200':
          description: |
            Success: Objective and project retrieved successfully. Possible messages:
            - "Success: Task's objective and project was retrieved."
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  task:
                    $ref: '#/components/schemas/Task'
                  objective:
                    $ref: '#/components/schemas/Objective'
                  project:
                    $ref: '#/components/schemas/Project'
        '401':
          description: |
            Unauthorized. Possible authentication errors from decorators:
            - "Request is missing access token. Please login to refresh access token"
            - "Deserialisation or Signiture verificaiton of access token cookie has failed! Reason: {error details}"
            - "Invalid Access Token! Reason: {error details}"
            - "Failure: User is not logged in (no b_sc). Please login!"
            - "Failure: Could not decrypt the bespoke_session cookies. Reason: {error details}"
            - "Failure: Refresh token is invalid. Please login"
            - "Failure: Please login. Refresh token has expired."
        '403':
          description: |
            Forbidden. Possible messages:
            - "Failure: The requested task does not belong to the user. Choose another one."
        '404':
          description: |
            Not found. Possible messages:
            - "Failure: The requested task is not in the database. Choose another one."
            - "Failure: User is not logged in (no rt). Please login!"
...
components:
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: integer
        projectNumber:
          type: integer
        type:
          type: string
        title:
          type: string
        description:
          type: string
        last_updated:
          type: string
          format: date-time
        tag:
          type: string
        user_id:
          type: integer
    Objective:
      type: object
      properties:
        id:
          type: integer
        objectiveNumber:
          type: integer
        status:
          type: string
        type:
          type: string
        title:
          type: string
        description:
          type: string
        deadline:
          type: string
          format: date-time
        scheduled_start:
          type: string
          format: date-time
        scheduled_finish:
          type: string
          format: date-time
        last_updated:
          type: string
          format: date-time
        tag:
          type: string
        project_id:
          type: integer
    Task:
      type: object
      properties:
        id:
          type: integer
        task_number:
          type: integer
        status:
          type: string
        description:
          type: string
        duration_est:
          type: integer
        duration:
          type: integer
        priority_score:
          type: integer
        scheduled_start:
          type: string
          format: date-time
        start:
          type: string
          format: date-time
        finish:
          type: string
          format: date-time
        is_recurring:
          type: boolean
        last_updated:
          type: string
          format: date-time
        was_paused:
          type: boolean
        parent_task_id:
          type: integer
        tag:
          type: string
        objective_id:
          type: integer
...